requirements.txt
flask>=2.3.0
pandas>=2.0.0
openpyxl>=3.1.0
bokeh>=3.2.0
pyyaml>=6.0
numpy>=1.24.0
scipy>=1.11.0
